Se implementan el mcu y mas. pendiente modificar el mas para q cada vble reciba el anchor de acuerdo a su posición

+++ Juego2D/Distractor.h
@@ -12,7 +12,7 @@ class Distractor: public QObject, public QGraphicsPixmapItem{
 
 public:
     Distractor(QGraphicsItem *parent = 0); //EL constructor por defecto asigna algo y evitamos posibles errores
-    int radio;
+    float radio;
     float vel;
 
     float radians = 0;
